Updated join room
