+++ backend/models/roomModel.js
@@ -1,28 +1,27 @@
-const db = require('../configs/db.config');
-
+const db = require('../configs/db_config');
 
 const Room = {
-    // Get room by invite code
+    // Get a room by invite code
     async getByInviteCode(invite_code) {
-        const query = 'SELECT room_id, room_name, invite_code FROM rooms WHERE invite_code = $1';
-        const { rows } = await db.query(query, [invite_code]); 
-        return rows[0] || null; // Return first room found or null if not found
+        const query = 'SELECT room_id, room_name FROM rooms WHERE invite_code = $1';
+        const { rows } = await db.query(query, [invite_code]);
+        return rows[0] || null; // Return the room if found, otherwise null
     },
 
     // Check if a user is already in the room
     async checkUserInRoom(user_id, room_id) {
         const query = 'SELECT * FROM room_participants WHERE user_id = $1 AND room_id = $2';
-        const { rows } = await db.query(query, [user_id, room_id]); // Query the room_participants table
-        return rows.length > 0; // Return true if user is found, false otherwise
+        const { rows } = await db.query(query, [user_id, room_id]);
+        return rows.length > 0; // Return true if the user is found in the room
     },
 
-    // Add a user to the room and create a notification
+    // Add a user to a room and create a notification
     async addUserToRoom(user_id, room_id, room_name) {
-        const client = await db.connect(); 
+        const client = await db.connect(); // Connect to the database
         try {
-            await client.query('BEGIN'); 
+            await client.query('BEGIN'); // Start a transaction
 
-            // Insert user into the room_participants table
+            // Insert the user into the room participants table
             await client.query(
                 'INSERT INTO room_participants (user_id, room_id, joined_at) VALUES ($1, $2, NOW())',
                 [user_id, room_id]
@@ -37,17 +36,15 @@ const Room = {
                 [user_id, message]
             );
 
-            await client.query('COMMIT'); // Commit the transaction if everything is successful
+            await client.query('COMMIT'); // Commit the transaction if successful
             return true;
         } catch (error) {
-            await client.query('ROLLBACK'); // Rollback the transaction in case of an error
-            console.error('Error adding user to room:', error);
-            throw new Error('Failed to add user to room or send notification');
+            await client.query('ROLLBACK'); // Rollback transaction in case of an error
+            throw error;
         } finally {
             client.release(); 
         }
     }
 };
 
-// Export the Room model
 module.exports = Room;
