--- backend/routes/joinRoomRoutes.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const roomController = require('../controllers/roomController');
-
-// Join room route
-router.post('/join-room', roomController.joinRoom);
-
-module.exports = router;
