+++ backend/server.js
@@ -1,11 +1,11 @@
 const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
-const session = require('express-session');
-const joinRoomRoutes = require('./routes/joinRoomRoutes');
-const authMiddleware = require('./middleware/authMiddleware');
+
 const db = require('./configs/db_config')
 const port = 3000;
+
+const roomRoutes = require('./routes/roomRoutes');
 
 app.use(express.json());
 
@@ -14,14 +14,6 @@ app.use(express.static(path.join(__dirname, '../frontend/loginpage')));
 
 const cors = require('cors');
 app.use(cors());
-
-// Middleware for session 
-app.use(session({ 
-    secret: process.env.SESSION_SECRET || 'secret_key', 
-    resave: false, 
-    saveUninitialized: false,
-    cookie: { secure: process.env.NODE_ENV === 'production' }
-}));
 
 
 app.get('/login', (req, res) => {
@@ -54,8 +46,6 @@ app.post('/users', db.createUser)
 app.put('/users/:id', db.updateUser)
 app.delete('/users/:id', db.deleteUser)
 
-
-// Register routes
-app.use('/api/rooms', authMiddleware, joinRoomRoutes);
+app.use('/api/rooms', roomRoutes);
 
 app.listen(port, () => console.log(`app listening on port ${port}`));
